+++ http/authentication_middleware.go
@@ -18,6 +18,7 @@ type AuthenticationHandler struct {
 
 	AuthorizationService platform.AuthorizationService
 	SessionService       platform.SessionService
+	SessionRenewDisabled bool
 
 	// This is only really used for it's lookup method the specific http
 	// handler used to register routes does not matter.
@@ -123,10 +124,12 @@ func (h *AuthenticationHandler) extractSession(ctx context.Context, r *http.Requ
 		return ctx, e
 	}
 
-	// if the session is not expired, renew the session
-	e = h.SessionService.RenewSession(ctx, s, time.Now().Add(platform.RenewSessionTime))
-	if e != nil {
-		return ctx, e
+	if !h.SessionRenewDisabled {
+		// if the session is not expired, renew the session
+		e = h.SessionService.RenewSession(ctx, s, time.Now().Add(platform.RenewSessionTime))
+		if e != nil {
+			return ctx, e
+		}
 	}
 
 	return platcontext.SetAuthorizer(ctx, s), nil