+++ kv/kvlog.go
@@ -108,7 +108,10 @@ func (s *Service) initializeKVLog(ctx context.Context, tx Tx) error {
 	return nil
 }
 
-var errKeyValueLogBoundsNotFound = fmt.Errorf("oplog not found")
+var errKeyValueLogBoundsNotFound = &platform.Error{
+	Code: platform.ENotFound,
+	Msg:  "oplog not found",
+}
 
 func (s *Service) getKeyValueLogBounds(ctx context.Context, tx Tx, key []byte) (*keyValueLogBounds, error) {
 	k := encodeKeyValueIndexKey(key)
@@ -381,7 +384,7 @@ func (s *Service) LastLogEntry(ctx context.Context, k []byte) ([]byte, time.Time
 func (s *Service) firstLogEntry(ctx context.Context, tx Tx, k []byte) ([]byte, time.Time, error) {
 	bounds, err := s.getKeyValueLogBounds(ctx, tx, k)
 	if err != nil {
-		return nil, bounds.StartTime(), err
+		return nil, time.Time{}, err
 	}
 
 	return s.getLogEntry(ctx, tx, k, bounds.StartTime())
@@ -390,7 +393,7 @@ func (s *Service) firstLogEntry(ctx context.Context, tx Tx, k []byte) ([]byte, t
 func (s *Service) lastLogEntry(ctx context.Context, tx Tx, k []byte) ([]byte, time.Time, error) {
 	bounds, err := s.getKeyValueLogBounds(ctx, tx, k)
 	if err != nil {
-		return nil, bounds.StopTime(), err
+		return nil, time.Time{}, err
 	}
 
 	return s.getLogEntry(ctx, tx, k, bounds.StopTime())