+++ kv/service.go
@@ -22,24 +22,37 @@ const OpPrefix = "kv/"
 type Service struct {
 	kv     Store
 	Logger *zap.Logger
+	Config ServiceConfig
 
 	IDGenerator    influxdb.IDGenerator
 	TokenGenerator influxdb.TokenGenerator
-	Hash           Crypt
-
-	time func() time.Time
+	influxdb.TimeGenerator
+	Hash Crypt
 }
 
 // NewService returns an instance of a Service.
-func NewService(kv Store) *Service {
-	return &Service{
+func NewService(kv Store, configs ...ServiceConfig) *Service {
+	s := &Service{
 		Logger:         zap.NewNop(),
 		IDGenerator:    snowflake.NewIDGenerator(),
 		TokenGenerator: rand.NewTokenGenerator(64),
 		Hash:           &Bcrypt{},
 		kv:             kv,
-		time:           time.Now,
+		TimeGenerator:  influxdb.RealTimeGenerator{},
 	}
+
+	if len(configs) > 0 {
+		s.Config = configs[0]
+	} else {
+		s.Config.SessionLength = influxdb.DefaultSessionLength
+	}
+
+	return s
+}
+
+// ServiceConfig allows us to configure Services
+type ServiceConfig struct {
+	SessionLength time.Duration
 }
 
 // Initialize creates Buckets needed.
@@ -115,12 +128,6 @@ func (s *Service) Initialize(ctx context.Context) error {
 
 		return s.initializeUsers(ctx, tx)
 	})
-}
-
-// WithTime sets the function for computing the current time. Used for updating meta data
-// about objects stored. Should only be used in tests for mocking.
-func (s *Service) WithTime(fn func() time.Time) {
-	s.time = fn
 }
 
 // WithStore sets kv store for the service.