+++ ui/src/templates/constants/defaultTemplates.ts
@@ -1,16 +1,17 @@
-import SYSTEM_TEMPLATE from 'src/templates/constants/system.json'
-import GETTING_STARTED from 'src/templates/constants/gettingStarted.json'
-
-import LOCAL_METRICS from 'src/templates/constants/localMetrics.json'
+import {
+  system,
+  localMetrics,
+  gettingStarted,
+} from '@influxdata/influxdb-templates'
 
 export const localMetricsTemplate = () => {
-  return LOCAL_METRICS
+  return localMetrics
 }
 
 export const systemTemplate = () => {
-  return SYSTEM_TEMPLATE
+  return system
 }
 
 export const gettingStartedWithFluxTemplate = () => {
-  return GETTING_STARTED
+  return gettingStarted
 }